+++ src/main/java/roomescape/Repository/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.Repository;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -11,8 +11,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.BaseRepository;
-import roomescape.reservationtime.ReservationTime;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservationtime.ReservationTime;
 
 @Repository
 public class ReservationRepository implements BaseRepository<Reservation> {
