+++ src/main/java/roomescape/Repository/ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservationtime;
+package roomescape.Repository;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -10,7 +10,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.BaseRepository;
+import roomescape.domain.reservationtime.ReservationTime;
 
 @Repository
 public class ReservationTimeRepository implements BaseRepository<ReservationTime> {
