+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,4 +1,4 @@
-package roomescape.reservationtime;
+package roomescape.controller;
 
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -10,6 +10,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import roomescape.domain.reservationtime.ReservationTime;
+import roomescape.service.ReservationTimeService;
 
 @Controller
 @RequestMapping("/times")
