+++ src/main/java/roomescape/domain/reservation/CreateReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.domain.reservation;
 
 import java.time.LocalDate;
