+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,10 +1,13 @@
-package roomescape.reservation;
+package roomescape.service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
-import roomescape.reservationtime.ReservationTime;
-import roomescape.reservationtime.ReservationTimeRepository;
+import roomescape.Repository.ReservationRepository;
+import roomescape.Repository.ReservationTimeRepository;
+import roomescape.domain.reservation.CreateReservationRequest;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservationtime.ReservationTime;
 
 @Service
 public class ReservationService {
